APContact stuff is moved to an own class
